delete view tests
