+++ lits/app/helpers/events_helper.rb
@@ -30,16 +30,4 @@ module EventsHelper
         event_time(event)
     end
   end
-
-  def tag_links(tags)
-    tags.split(',').map { |tag| link_to tag.strip, tag_path(tag.strip) }.join(', ')
-  end
-
-  def tag_cloud(tags, classes)
-    max = tags.sort_by(&:count).last
-    tags.sort_by(&:name).each do |tag|
-      index = tag.count.to_f / max.count * (classes.size - 1)
-      yield(tag, classes[index.round])
-    end
-  end
 end
