+++ lits/app/helpers/tags_helper.rb
@@ -1,2 +1,13 @@
 module TagsHelper
+  def tag_links(tags)
+    tags.split(',').map { |tag| link_to tag.strip, tag_path(tag.strip) }.join(', ')
+  end
+
+  def tag_cloud(tags, classes)
+    max = tags.sort_by(&:count).last
+    tags.sort_by(&:name).each do |tag|
+      index = tag.count.to_f / max.count * (classes.size - 1)
+      yield(tag, classes[index.round])
+    end
+  end
 end
