--- lits/spec/views/tags/edit.html.erb_spec.rb
@@ -1,21 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "tags/edit", type: :view do
-  before(:each) do
-    @tag = assign(:tag, Tag.create!(
-      :name => "MyString",
-      :slug => "MyString"
-    ))
-  end
-
-  it "renders the edit tag form" do
-    render
-
-    assert_select "form[action=?][method=?]", tag_path(@tag), "post" do
-
-      assert_select "input#tag_name[name=?]", "tag[name]"
-
-      assert_select "input#tag_slug[name=?]", "tag[slug]"
-    end
-  end
-end
