--- lits/spec/views/tags/index.html.erb_spec.rb
@@ -1,22 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "tags/index", type: :view do
-  before(:each) do
-    assign(:tags, [
-      Tag.create!(
-        :name => "Name",
-        :slug => "Slug"
-      ),
-      Tag.create!(
-        :name => "Name",
-        :slug => "Slug"
-      )
-    ])
-  end
-
-  it "renders a list of tags" do
-    render
-    assert_select "tr>td", :text => "Name".to_s, :count => 2
-    assert_select "tr>td", :text => "Slug".to_s, :count => 2
-  end
-end
