--- lits/spec/views/tags/new.html.erb_spec.rb
@@ -1,21 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "tags/new", type: :view do
-  before(:each) do
-    assign(:tag, Tag.new(
-      :name => "MyString",
-      :slug => "MyString"
-    ))
-  end
-
-  it "renders new tag form" do
-    render
-
-    assert_select "form[action=?][method=?]", tags_path, "post" do
-
-      assert_select "input#tag_name[name=?]", "tag[name]"
-
-      assert_select "input#tag_slug[name=?]", "tag[slug]"
-    end
-  end
-end
