--- lits/spec/views/tags/show.html.erb_spec.rb
@@ -1,16 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "tags/show", type: :view do
-  before(:each) do
-    @tag = assign(:tag, Tag.create!(
-      :name => "Name",
-      :slug => "Slug"
-    ))
-  end
-
-  it "renders attributes in <p>" do
-    render
-    expect(rendered).to match(/Name/)
-    expect(rendered).to match(/Slug/)
-  end
-end
